Refactor SBOM generation to use nox session output
