+++ .github/workflows/build-and-publish.yml
@@ -34,13 +34,10 @@ jobs:
 
       - name: Generate SBOM
         id: generate-sbom
-        run: poetry run -- nox -s dependency:sbom
-      - name: Rename SBOM
-        id: rename-sbom
         run: |
-          SBOM_NAME=$(poetry version --no-ansi | awk '{gsub(/-/, "_", $1); print $1"-"$2".spdx.json"}')
-          mv bom.spdx.json $SBOM_NAME
-          echo "SBOM_NAME=$SBOM_NAME" >> $GITHUB_ENV
+          SBOM_NAME="$(poetry run -- nox -s dependency:sbom)"
+          echo "SBOM_NAME=$SBOM_NAME" >> "$GITHUB_OUTPUT"
+
       - name: Publish Release to PyPi
         id: publish-release-to-pypi
         env:
@@ -52,9 +49,10 @@ jobs:
         id: publish-release-to-github
         env:
           GITHUB_TOKEN: ${{ secrets.GITHUB_TOKEN }}
+          SBOM_NAME: ${{ steps.generate-sbom.outputs.SBOM_NAME }}
         run: >-
           gh release create ${GITHUB_REF_NAME}
           --title ${GITHUB_REF_NAME}
           --notes-file ./doc/changes/changes_${GITHUB_REF_NAME}.md
           dist/*
-          "${{ env.SBOM_NAME }}"
+          "$SBOM_NAME"
